Created predictable shuffling for column values
CoreFunctions.mappedColumnShuffle guarantees the same return value
for a given column value.

To achieve that, the reflection used to call a CoreFunction method
had to be updated so it's more strongly-typed.  Now parameters are
matched from the requirement file's Function tag by name, and the
type must either be unspecified or match the parameter definition
for the method in CoreFunctions.

Special values can be passed... this functionality may need some
cleanup to make it nicer and more predictable. At the moment a
parameter named 'value' on a CoreFunction method would be passed
the column's value.  A ResultSet 'row' parameter would be passed
the current row.