Added NotLike and NotEquals "Exclude" rules
NotLike and NotEquals act as "Inclusion" rules.  A column's value
must match at least one NotEquals or NotLike rules to be included.

Column definitions in Requirements file may repeat the name of a
single column, so different anonymization rules can be applied
based on the exclusion/inclusion of column values.