Fix OutOfMemoryException for large MySQL tables
The fix forces the MySQL driver into 'streaming' mode so the entire
RS isn't stored in memory.  Unfortunately in streaming mode update
queries on the same connection fail, and so a separate connection to
the database seems to be the best way around it (skipping rows also
doesn't work, and the RS cannot be updatable in streaming mode).