Option to not calculate Score - Performance issue when running select count(*) over large tables. New parameter score_calculation is needed
Found some nasty performance issue when running score calculation for
very large tables as it needs (as it is now) rowcount. Running select
count(*) against very large tables is not a good ideia. Option is now
to avoid score calculation. For that score_calculation parameters is
added. Beside that, The “analyzing table” output was improved with
column name, instead of repeating the table name.