Added Recursive File Discovery against directories specified in directories parameter
Decided to implement Recursive file discovery against the directories
specified in the parameter in filediscovery.properties. If you have a
repository (windows shares) with huge amounts of directories and
subdirectories, it is much faster to do it automatically.
Also cleaned some output.
Please review if that feature makes sense for you.